requirements.txt
anyio==3.6.1
asgiref==3.5.2
boto3==1.23.5
botocore==1.26.5
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
cycler==0.11.0
fastapi==0.78.0
fonttools==4.33.3
h11==0.13.0
idna==3.3
jmespath==1.0.0
kiwisolver==1.4.2
matplotlib==3.5.2
numpy==1.22.4
packaging==21.3
pandas==1.4.2
Pillow==9.1.1
psycopg2-binary==2.9.3
pydantic==1.9.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
python-multipart==0.0.5
pytz==2022.1
requests==2.27.1
s3transfer==0.5.2
scipy==1.8.1
seaborn==0.11.2
six==1.16.0
sniffio==1.2.0
spotipy==2.19.0
starlette==0.19.1
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
